Move icons to dedicated file
