--- src/accordion-item/components/icon.js
@@ -1,19 +1,0 @@
-import { SVG, Path } from '@wordpress/components';
-
-const Icon = (
-	<SVG
-		fill="none"
-		viewBox="0 0 24 24"
-		xmlns="http://www.w3.org/2000/svg"
-		width="24"
-		height="24"
-	>
-		<Path
-			d="m19 9c0-.27614-.2239-.5-.5-.5h-13c-.27614 0-.5.22386-.5.5v7.5c0 .2761.22386.5.5.5h13c.2761 0 .5-.2239.5-.5zm-13.875-3.5c-.89746 0-1.625.72754-1.625 1.625v9.75c0 .8975.72754 1.625 1.625 1.625h13.75c.8975 0 1.625-.7275 1.625-1.625v-9.75c0-.89746-.7275-1.625-1.625-1.625z"
-			fill="currentColor"
-			fillRule="evenodd"
-		></Path>
-	</SVG>
-);
-
-export default Icon;
