+++ src/accordion/index.js
@@ -20,7 +20,7 @@ import './style.scss';
 import json from './block.json';
 import Edit from './edit';
 import Save from './save';
-import Icon from './components/icon';
+import { Accordion as Icon } from '../components/icons';
 
 const { name } = json;
 
@@ -29,7 +29,7 @@ const { name } = json;
  *
  * @see https://developer.wordpress.org/block-editor/reference-guides/block-api/block-registration/
  */
-registerBlockType( name, {
+registerBlockType(name, {
 	...json,
 
 	/**
@@ -46,4 +46,4 @@ registerBlockType( name, {
 	 * @see ./save.js
 	 */
 	save: Save,
-} );
+});
